extend second operand w/ arithmetic flag setting to uint64_t (stops overflow when checking cases such as ADC/SBC). fuzzarm passes

+++ agbe/ARM7TDMI.h
@@ -123,5 +123,5 @@ private:
 
 	//Flag setting
 	void setLogicalFlags(uint32_t result, int carry);
-	void setArithmeticFlags(uint32_t input, uint32_t operand, uint32_t result, bool addition);
+	void setArithmeticFlags(uint32_t input, uint64_t operand, uint32_t result, bool addition);
 };
